# Conflicts:
#	mercury/src/main/db/athenadb.changelog-1.3.xml
#	mercury/src/main/java/org/broadinstitute/gpinformatics/athena/boundary/billing/QuoteImportInfo.java
#	mercury/src/main/java/org/broadinstitute/gpinformatics/athena/entity/orders/ProductOrderSample.java
#	mercury/src/main/java/org/broadinstitute/gpinformatics/athena/presentation/orders/BillingLedgerActionBean.java
#	mercury/src/main/java/org/broadinstitute/gpinformatics/infrastructure/test/dbfree/ProductOrderSampleTestFactory.java
#	mercury/src/test/java/org/broadinstitute/gpinformatics/athena/entity/orders/PDOSampleBilledStatusTest.java
#	mercury/src/test/java/org/broadinstitute/gpinformatics/athena/entity/orders/ProductOrderSampleTest.java
#	mercury/src/test/java/org/broadinstitute/gpinformatics/infrastructure/sap/SapIntegrationServiceImplDBFreeTest.java

+++ mercury/src/test/java/org/broadinstitute/gpinformatics/athena/entity/orders/PDOSampleBilledStatusTest.java
@@ -11,6 +11,7 @@ import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
+import java.math.BigDecimal;
 import java.text.MessageFormat;
 import java.util.Date;
 
@@ -37,9 +38,9 @@ public class PDOSampleBilledStatusTest {
         pdo.addSample(pdoSample);
         pdo.getProduct().setPrimaryPriceItem(primaryPriceItem);
         if(pdo.hasSapQuote()) {
-            pdoSample.addLedgerItem(new Date(System.currentTimeMillis()), pdo.getProduct(),3d, null);
+            pdoSample.addLedgerItem(new Date(System.currentTimeMillis()), pdo.getProduct(), BigDecimal.valueOf(3), null);
         } else {
-            pdoSample.addLedgerItem(new Date(System.currentTimeMillis()), primaryPriceItem,3d);
+            pdoSample.addLedgerItem(new Date(System.currentTimeMillis()), primaryPriceItem,BigDecimal.valueOf(3));
         }
 
         BillingSession billingSession = new BillingSession(3L,pdoSample.getLedgerItems());
